wip: Moved backend into seperate modules
Now only functions that are to be called across the IPC boundary are
imported. Should make compartmentalizing responsibilities easier.
